requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
Pillow==8.4.0
sqlparse==0.4.2
